audit: add changes for audit NM-0544-A (#69)
* fix(appchain): use integrity utils to verify the fact hash

By using is_fact_hash_valid_with_security from the integrity utils,
all the verifications are checked to avoid front-running and locking piltover that was only checking the first one.

* fix(state): add missing input validation

To be consistent with the L1 implementation of Starknet,
two checks have been added to validate the inputs of the update function:

1. The previous block hash in the SNOS output must match the current block hash of the state
2. The block number must always be greater than the current state (without requirement to be +1)

* fix: remove KZG DA support in output deserialization

KZG DA when supported yields a different Starknet OS output,
which doesn't contain the state diff.

And the KZG proofs are also expected as input.This is currently not supported by piltover.
At the moment, since no work has been done to fully support KZG DA,
the support is disabled by panicking if the flag is set to true in the Starknet OS output.

* fix: remove full output in output deserialization

Due to the current design of piltover, the full output
is never expected to reach piltover.
Offchain components are using the RPC to fetch block n-1
data in order to prove the state transition.

* fix: remove unused code

* fix(messaging): ensure message cancellation timer is not reset

To cancel a message, a cancellation request must first be sent (start_message_cancellation),
which uses the block timestamp to then track the necessary delay before cancel_message can be called.

The change here ensures that this timer is not reset if the start_message_cancellation entrypoint is called multiple times.

* fix: enforce onchain_data_hash and onchain_data_size values

Since those values usage is not well defined in piltover compared to what
Starknet does on Ethereum, those values are now fixed by piltover to 0.

This avoids having those values passed as argument by the operator,
which could be subject to manipulation of the state transaction fact.

* fix: add validation of the bootloader hash in layout bridge output

Layout bridge is a verifier written in Cairo.
To ensure that the program that has been verified is legitimate,
and add consistency with the computed fact, the layout bridge output
that contains the executed program is now verified.

In the current implementation, this program hash is the bootloader hash,
since Starknet OS execution is bootloaded. And this is the trace
being verified by the layout bridge.

* fix: generate bindings

* fix(snos_output): add check for the os_program_hash

* refacto(messaging): add Cancelling variant to track message state

Adding a variant to track the cancelling state of a message
could be helpful when a message status is queried by external users.

Also, the new variant ensures that the cancellation can't be done twice.

* fix: ensure bindgen is correctly done

* fix: fix typo in comment